chatbot/requirements.txt
greenlet==3.1.1
fastapi==0.116.1
langchain-community==0.3.31
langchain-openai==0.3.35
langchain-huggingface==0.3.1
python-dotenv>=1.0.1
pypdf>=4.0.0
uvicorn==0.35.0
